refactor(mapgen): centralize mountain strategies in morphology domain
Move all mountain-related logic from scattered orogeny/narrative locations
to a unified domain/morphology/mountains module:

- Create MountainBeltStrategy.ts consolidating belt detection and wind tagging
- Move MountainPassStrategy.ts from corridors to morphology
- Create unified cache.ts for mountain/orogeny cache management
- Update mountains/index.ts with comprehensive exports
- Convert old narrative/orogeny files to backward-compat re-exports
- Update pipeline steps to import from new centralized location

This centralizes mountain terrain logic where it belongs (morphology) while
maintaining backward compatibility through re-exports.
