+++ packages/mapgen-core/src/domain/narrative/orogeny/cache.ts
@@ -1,34 +1,10 @@
-import type { ExtendedMapContext } from "../../../core/types.js";
-
-export interface OrogenyCacheInstance {
-  belts: Set<string>;
-  windward: Set<string>;
-  lee: Set<string>;
-}
-
-const OROGENY_CACHE_ARTIFACT_KEY = "story:orogenyCache";
-
-function createCache(): OrogenyCacheInstance {
-  return { belts: new Set(), windward: new Set(), lee: new Set() };
-}
-
-export function getOrogenyCache(ctx: ExtendedMapContext | null | undefined): OrogenyCacheInstance {
-  if (!ctx) return createCache();
-  const existing = ctx.artifacts?.get(OROGENY_CACHE_ARTIFACT_KEY) as OrogenyCacheInstance | undefined;
-  if (existing) return existing;
-  const created = createCache();
-  ctx.artifacts?.set(OROGENY_CACHE_ARTIFACT_KEY, created);
-  return created;
-}
-
-export function resetOrogenyCache(ctx: ExtendedMapContext | null | undefined): void {
-  if (!ctx) return;
-  ctx.artifacts?.set(OROGENY_CACHE_ARTIFACT_KEY, createCache());
-}
-
-export function clearOrogenyCache(ctx: ExtendedMapContext | null | undefined): void {
-  const cache = getOrogenyCache(ctx);
-  cache.belts.clear();
-  cache.windward.clear();
-  cache.lee.clear();
-}
+/**
+ * Orogeny Cache — Backward compatibility re-export
+ * @deprecated Import from "../../morphology/mountains" instead
+ */
+export {
+  getMountainCache as getOrogenyCache,
+  resetMountainCache as resetOrogenyCache,
+  clearMountainCache as clearOrogenyCache,
+  type MountainCacheInstance as OrogenyCacheInstance,
+} from "../../morphology/mountains/index.js";
