+++ packages/mapgen-core/src/domain/narrative/orogeny/index.ts
@@ -1,3 +1,22 @@
-export * from "./cache.js";
-export * from "./wind.js";
-export * from "./belts.js";
+/**
+ * Orogeny Re-exports — Backward compatibility shim
+ *
+ * Mountain belt detection has been centralized in domain/morphology/mountains.
+ * This module re-exports for backward compatibility with existing code.
+ *
+ * @deprecated Import from "../../morphology/mountains" instead
+ */
+
+// Re-export cache utilities
+export {
+  getMountainCache as getOrogenyCache,
+  resetMountainCache as resetOrogenyCache,
+  clearMountainCache as clearOrogenyCache,
+  type MountainCacheInstance as OrogenyCacheInstance,
+} from "../../morphology/mountains/index.js";
+
+// Re-export belt detection
+export {
+  detectMountainBelts as storyTagOrogenyBelts,
+  computeZonalWindStep as zonalWindStep,
+} from "../../morphology/mountains/index.js";
