+++ packages/mapgen-core/src/pipeline/narrative/StoryOrogenyStep.ts
@@ -1,6 +1,6 @@
 import type { ExtendedMapContext } from "../../core/types.js";
 import { M3_STANDARD_STAGE_PHASE, type MapGenStep } from "../index.js";
-import { storyTagOrogenyBelts } from "../../domain/narrative/orogeny/index.js";
+import { detectMountainBelts } from "../../domain/morphology/mountains/index.js";
 
 export interface StoryOrogenyStepOptions {
   requires: readonly string[];
@@ -18,7 +18,7 @@ export function createStoryOrogenyStep(
     provides: options.provides,
     shouldRun: options.shouldRun ? () => options.shouldRun?.() === true : undefined,
     run: (context) => {
-      storyTagOrogenyBelts(context);
+      detectMountainBelts(context);
     },
   };
 }
